Add timer.h, formatted_log.h, remove delays

+++ src/I2C_Im_Slave.cpp
@@ -3,18 +3,18 @@
 /* -------------------------- PRIVATE --------------------------------*/
 int I2C_Im_Slave::device_counter = 0;
 
-bool I2C_Im_Slave::check_synchronizing(int message) {
+bool I2C_Im_Slave::check_synchronizing(uint32_t message) {
 	this->synchronizing = (message == I2C_Device::SYNC_MESSAGE);
 	
 	return this->synchronizing;
 }
 
 /* -------------------------- PUBLIC --------------------------------*/
-int I2C_Im_Slave::get_request_from_master() {
+uint32_t I2C_Im_Slave::get_request_from_master() {
 	return this->request_from_master;
 }
 
-void I2C_Im_Slave::set_request_from_master(int request) {
+void I2C_Im_Slave::set_request_from_master(uint32_t request) {
 	this->request_from_master = request;
 }
 
@@ -24,7 +24,7 @@ bool I2C_Im_Slave::is_synchronizing() {
 
 void I2C_Im_Slave::receive_message() {
 	while (0 < Wire.available()) {
-		int x = this->get_u_int_from_response();
+		uint8_t x = this->get_u_int_from_response();
 
 		this->set_request_from_master(x);
 		this->check_synchronizing(x);
